feat: EPUB handling: find content.opf files
With proper configured FAT filesystem, the actual EPUB handling started.
The commonly content.opf called file is the rootfile of an EPUB,
containing information like title, author, and references to the
single chapter files in reading order.

This adds a basic, efficient search for this root documents with
the assumption the rootfile is "OEBPS/content.opf".

Valid root files at other locations or with different names are
not found right now!
Later, the rootfile should be found by parsing the required
"META-INF/container.xml" file which must contain the correct path.

Signed-off-by: Anna-Lena Marx <anna-lena.marx@web.de>