feat: epub parsing: Get titles and authors of all books, get next and prev page
Adds support for parsing titles and authors from the metadata file for all
book directories located on the SD card (with expected name right now) and
store the related information for each book during runtime.

Also adds support for getting the basic, unparsed content of an actual
book chapter (x)html file and moving through this file for- and backwards.

Signed-off-by: Anna-Lena Marx <anna-lena.marx@web.de>