fix/cleanup: move implementations back into the app directory (#19)
The static linking for libraries in a lib folder using the zephyr_library()
macro is rather flaky and ordering dependend.
Issues and failures are not visible during compile time and the application's
behaviour changes depending on the include ordering in CMakeLists.txt.
As a result, the reordering is switched back for now until there is a
better and more reliable way to modularize it.

Fixes #18